Pruebas en NUnit, cambio en los mensajes de alerta.

+++ floridaimportdotnet/ClientApp/src/app/florida/cliente-eliminar/cliente-eliminar.component.ts
@@ -4,6 +4,8 @@ import { ClienteService } from 'src/app/services/cliente.service';
 import { ActivatedRoute } from '@angular/router';
 import { AlertModalComponent } from 'src/app/@base/alert-modal/alert-modal.component';
 import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { AlertModalEliminarComponent } from '../../@base/alert-modal-eliminar/alert-modal-eliminar.component';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-cliente-eliminar',
@@ -13,14 +15,21 @@ import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
 export class ClienteEliminarComponent implements OnInit {
 
   cliente: Cliente;
-  constructor(private clienteService: ClienteService, private rutaActiva: ActivatedRoute, private modalService: NgbModal) { }
+    rol: string;
+  constructor(private clienteService: ClienteService,
+    private rutaActiva: ActivatedRoute,
+    private modalService: NgbModal) { }
 
   ngOnInit(): void {
-    this.cliente = new Cliente();
-    const correo = this.rutaActiva.snapshot.params.identificacion;
-    this.clienteService.getCorreo(correo).subscribe(c => {
-      this.cliente = c;
-    });
+    this.rol = sessionStorage.getItem('User');
+
+    if (this.rol == "Admin") {
+      this.cliente = new Cliente();
+      const correo = this.rutaActiva.snapshot.params.identificacion;
+      this.clienteService.getCorreo(correo).subscribe(c => {
+        this.cliente = c;
+      });
+    }
   }
 
   confDelete(){
@@ -28,13 +37,47 @@ export class ClienteEliminarComponent implements OnInit {
   }
   
   delete() {
-    this.clienteService.delete(this.cliente.correo).subscribe(c => {
-      const messageBox = this.modalService.open(AlertModalComponent)
 
-        messageBox.componentInstance.title = "Resultado de eliminacion de datos.";
-        messageBox.componentInstance.message = 'La persona '+ this.cliente.nombre+' fue eliminada correctamente.';
+    Swal.fire({
+      title: '¿Está seguro?',
+      text: "Si elimina a "+this.cliente.nombre+", sus datos no podrán ser recuperados!",
+      icon: 'warning',
+      showCancelButton: true,
+      cancelButtonColor: '#d33',
+      confirmButtonColor: '#22bb33',
+      cancelButtonText: 'cancelar!',
+      confirmButtonText: 'Eliminar!',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.clienteService.delete(this.cliente.correo).subscribe(c => {
 
-    });
+          Swal.fire({
+            title: 'Resultado de eliminación de datos!',
+            text: 'El cliente ' + this.cliente.nombre + ' ' + this.cliente.apellido + ' fué eliminada correctamente.',
+            icon: 'success',
+            confirmButtonColor: '#22bb33',
+          })
+          
+        });
+
+        window.location.href = "https://localhost:5001/clienteConsulta";
+
+        
+      } else if (
+        /* Read more about handling dismissals below */
+        result.dismiss === Swal.DismissReason.cancel
+      ) {
+        Swal.fire({
+          title: 'Cancelado',
+          text: 'La eliminación de ' + this.cliente.nombre + ' fué cancelada.',
+          icon: 'error',
+          confirmButtonColor: '#22bb33',
+        })
+      }
+    })
+
+
   }
 
 }
